wip: plan_for_grouping
