+++ crates/ndc-query-plan/src/query_plan/fields.rs
@@ -1,12 +1,10 @@
-use std::collections::BTreeMap;
-
 use derivative::Derivative;
 use indexmap::IndexMap;
 use ndc_models as ndc;
 
 use crate::Type;
 
-use super::{Aggregate, ConnectorTypes, Expression, RelationshipArgument};
+use super::{Aggregate, ConnectorTypes};
 
 #[derive(Derivative)]
 #[derivative(Clone(bound = ""), Debug(bound = ""), PartialEq(bound = ""))]
@@ -52,19 +50,4 @@ pub enum NestedField<T: ConnectorTypes> {
     Object(NestedObject<T>),
     Array(NestedArray<T>),
     // TODO: ENG-1464 add `Collection(NestedCollection)` variant
-}
-
-#[derive(Derivative)]
-#[derivative(Clone(bound = ""), Debug(bound = ""), PartialEq(bound = ""))]
-pub struct PathElement<T: ConnectorTypes> {
-    /// Path to a nested field within an object column that must be navigated
-    /// before the relationship is navigated.
-    /// Only non-empty if the 'relationships.nested' capability is supported.
-    pub field_path: Option<Vec<ndc::FieldName>>,
-    /// The name of the relationship to follow
-    pub relationship: ndc::RelationshipName,
-    /// Values to be provided to any collection arguments
-    pub arguments: BTreeMap<ndc::ArgumentName, RelationshipArgument<T>>,
-    /// A predicate expression to apply to the target collection
-    pub predicate: Option<Box<Expression<T>>>,
 }
